+++ src/cli/Agent/App/Install.php
@@ -15,7 +15,7 @@ class Install extends \HubletoMain\Cli\Agent\Command
       $this->cli->red("What app you want to install? Usage: php hubleto app install <APP_NAME>\n");
     }
 
-    require_once($this->main->configAsString('accountDir', __DIR__) . "/ConfigEnv.php");
+    require_once($this->main->config->getAsString('accountDir', __DIR__) . "/ConfigEnv.php");
 
     $appManager->installApp($appNamespace, [], $forceInstall);
     $this->cli->cyan("{$appNamespace} installed successfully.\n");