+++ src/cli/Agent/Project/GenerateDemoData.php
@@ -29,7 +29,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
     $idProfile = 1; // plati za predpokladu, ze tento command sa spusta hned po CommandInit
 
     $mUser = new \HubletoApp\Community\Settings\Models\User($this->main);
-    $idUserSalesManager = $mUser->recordManager->create([
+    $idUserSalesManager = $mUser->record->create([
       "first_name" => "Sales",
       "last_name" => "Manager",
       "email" => "test@user.sk",
@@ -40,13 +40,13 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
     ])['id'];
 
     $mUserHasRole = new \HubletoApp\Community\Settings\Models\UserHasRole($this->main);
-    $mUserHasRole->recordManager->create([
+    $mUserHasRole->record->create([
       "id_user" => $idUserSalesManager,
       "id_role" => \HubletoApp\Community\Settings\Models\UserRole::ROLE_SALES_MANAGER,
     ]);
 
     $mInvoiceProfile = new \HubletoApp\Community\Settings\Models\InvoiceProfile($this->main);
-    $mInvoiceProfile->recordManager->create(['name' => 'Test Profile 1']);
+    $mInvoiceProfile->record->create(['name' => 'Test Profile 1']);
 
     //Documents
     $mDocuments = new \HubletoApp\Community\Documents\Models\Document($this->main);
@@ -106,7 +106,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
   {
     $mInvoiceProfile = new \HubletoApp\Community\Settings\Models\InvoiceProfile($this->main);
     $mInvoiceProfile->install();
-    $mInvoiceProfile = $mInvoiceProfile->recordManager->create([
+    $mInvoiceProfile = $mInvoiceProfile->record->create([
       "name" => "Test Invoice Profile"
     ]);
   }
@@ -116,13 +116,13 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
 
     $mTag = new \HubletoApp\Community\Settings\Models\Tag($this->main);
 
-    $mTag->recordManager->create([
+    $mTag->record->create([
       'name' => "Category 1",
     ]);
-    $mTag->recordManager->create([
+    $mTag->record->create([
       'name' => "Category 2",
     ]);
-    $mTag->recordManager->create([
+    $mTag->record->create([
       'name' => "Category 3",
     ]);
   }
@@ -240,7 +240,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
     foreach ($customers as $customerCsvData) {
       $customer = explode(",", trim($customerCsvData));
 
-      $idCustomer = $mCustomer->recordManager->create([
+      $idCustomer = $mCustomer->record->create([
         "id_country" => $customer[0],
         "customer_id" => $customer[1],
         "name" => $customer[2],
@@ -259,7 +259,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
 
       $tagCount = rand(0, 3);
       for ($i = 0; $i < $tagCount; $i++) {
-        $mCustomerTag->recordManager->create([
+        $mCustomerTag->record->create([
           "id_customer" => $idCustomer,
           "id_tag" => rand(1, 3)
         ]);
@@ -430,7 +430,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
     ];
 
     foreach ($persons as $person) {
-      $idPerson = $mPerson->recordManager->create([
+      $idPerson = $mPerson->record->create([
         "id_customer" => rand(1, 100),
         "first_name" => $person[2],
         "last_name" => $person[3],
@@ -440,7 +440,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
         "date_created" => date("Y-m-d", rand(strtotime("-1 month"), strtotime("+1 month"))),
       ])['id'];
 
-      $mContact->recordManager->create([
+      $mContact->record->create([
         "id_person" => $idPerson,
         "type" => "email",
         "value" => str_replace("'", "", (string) iconv('UTF-8', 'ASCII//TRANSLIT', $person[7])),
@@ -448,7 +448,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
       ]);
 
       $phoneNumber = "+421 9" . rand(0, 3) . rand(4, 8) . " " . rand(0, 9) . rand(0, 9) . rand(0, 9) . " " . rand(0, 9) . rand(0, 9) . rand(0, 9);
-      $mContact->recordManager->create([
+      $mContact->record->create([
         "id_person" => $idPerson,
         "type" => "number",
         "value" => $phoneNumber,
@@ -457,13 +457,13 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
 
       $tagCount = rand(0, 3);
       for ($i = 0; $i < $tagCount; $i++) {
-        $mPersonTag->recordManager->create([
+        $mPersonTag->record->create([
           "id_person" => $idPerson,
           "id_tag" => rand(1, 3)
         ]);
       }
 
-      $mAddress->recordManager->create([
+      $mAddress->record->create([
         "id_person" => $idPerson,
         "street_line_1" => $streets[rand(0, 9)],
         "street_line_2" => "",
@@ -513,7 +513,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
   //           break;
   //       }
 
-  //       $activityId = $mActivity->recordManager->create([
+  //       $activityId = $mActivity->record->create([
   //         "id_activity_type" => $activityType,
   //         "subject" => $randomSubject,
   //         "date_start" => $date,
@@ -523,7 +523,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
   //         "id_person" => null,
   //       ]);
 
-  //       $mCustomerActivity->recordManager->create([
+  //       $mCustomerActivity->record->create([
   //         "id_customer" => $customer->id,
   //         "id_activity" => $activityId
   //       ]);
@@ -541,7 +541,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
 
     //Create all services
     foreach ($serviceNames as $serviceName) {
-      $mService->recordManager->create([
+      $mService->record->create([
         "name" => $serviceName,
         "price" => rand(10,100),
         "id_currency" => 1,
@@ -551,13 +551,13 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
     //Generate Billing Accounts and connect Services to them
     /* foreach ($customers as $customer) {
       $NOServices = rand(1, count($serviceNames));
-      $idBillingAccount = $mBillingAccount->recordManager->create([
+      $idBillingAccount = $mBillingAccount->record->create([
         "id_customer" => $customer->id,
         "description" => "Fakturácia",
       ]);
 
       for ($i = 0; $i < $NOServices; $i++) {
-        $mBillingAccountService->recordManager->create([
+        $mBillingAccountService->record->create([
           "id_billing_account" => $idBillingAccount,
           "id_service" => rand(1, count($serviceNames) - 1)
         ]);
@@ -565,13 +565,13 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
     }
     foreach ($customers as $customer) {
       $NOServices = rand(1, count($serviceNames));
-      $idBillingAccount = $mBillingAccount->recordManager->create([
+      $idBillingAccount = $mBillingAccount->record->create([
         "id_customer" => $customer->id,
         "description" => "Objednávky",
       ]);
 
       for ($i = 0; $i < $NOServices; $i++) {
-        $mBillingAccountService->recordManager->create([
+        $mBillingAccountService->record->create([
           "id_billing_account" => $idBillingAccount,
           "id_service" => rand(1, count($serviceNames) - 1)
         ]);
@@ -606,7 +606,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
       $leadDateCreated = date("Y-m-d", $leadDateCreatedTs);
       $leadDateClose = date("Y-m-d", $leadDateCreatedTs + rand(4, 6)*24*3600);
 
-      $idLead = $mLead->recordManager->create([
+      $idLead = $mLead->record->create([
         "identifier" => $identifierPrefixes[rand(0,2)] . rand(1,3000),
         "title" => $titles[rand(0, count($titles)-1)],
         "id_customer" => $customer->id,
@@ -621,13 +621,13 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
         "date_created" => $leadDateCreated,
       ])['id'];
 
-      $mLeadHistory->recordManager->create([
+      $mLeadHistory->record->create([
         "description" => "Lead created",
         "change_date" => date("Y-m-d", rand(strtotime("-1 month"), strtotime("+1 month"))),
         "id_lead" => $idLead
       ]);
 
-      $mLeadActivity->recordManager->create([
+      $mLeadActivity->record->create([
         "subject" => "Follow-up call",
         "date_start" => $leadDateCreated,
         "id_lead" => $idLead,
@@ -636,7 +636,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
         "id_user" => 1,
       ]);
 
-      $mLeadTag->recordManager->create([
+      $mLeadTag->record->create([
         "id_lead" => $idLead,
         "id_tag" => rand(1,3)
       ]);
@@ -666,7 +666,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
       $dealDateCreated = date("Y-m-d H:i:s", $dealDateCreatedTs);
       $dealDateClose = date("Y-m-d H:i:s", strtotime("+1 month", $dealDateCreatedTs));
 
-      $idDeal = $mDeal->recordManager->create([
+      $idDeal = $mDeal->record->create([
         "identifier" => $lead->identifier,
         "title" => $lead->title,
         "id_customer" => $lead->id_customer,
@@ -686,7 +686,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
         "date_progress_update" => $progress != 2 ? $dealDateClose : null,
       ])['id'];
 
-      $mLeadHistory->recordManager->create([
+      $mLeadHistory->record->create([
         "description" => "Converted to a deal",
         "change_date" => date("Y-m-d", rand(strtotime("-1 month"), strtotime("+1 month"))),
         "id_lead" => $lead->id
@@ -698,7 +698,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
       ;
 
       foreach ($leadHistories as $leadHistory) { // @phpstan-ignore-line
-        $mDealHistory->recordManager->create([
+        $mDealHistory->record->create([
           "description" => $leadHistory->description,
           "change_date" => $leadHistory->change_date,
           "id_deal" => $idDeal
@@ -710,7 +710,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
         ->first()
       ;
 
-      $mDealActivity->recordManager->create([
+      $mDealActivity->record->create([
         "subject" => "Follow-up call",
         "date_start" => $dealDateCreated,
         "id_deal" => $idDeal,
@@ -719,7 +719,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
         "id_user" => 1,
       ]);
 
-      $mDealTag->recordManager->create([
+      $mDealTag->record->create([
         "id_deal" => $idDeal,
         "id_tag" => $leadTag->id_tag
       ]);
@@ -732,32 +732,32 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
     \HubletoApp\Community\Products\Models\Supplier $mSupplier,
   ): void {
 
-    $mGroup->recordManager->create([
+    $mGroup->record->create([
       "title" => "Food"
     ]);
-    $mGroup->recordManager->create([
+    $mGroup->record->create([
       "title" => "Furniture"
     ]);
-    $mGroup->recordManager->create([
+    $mGroup->record->create([
       "title" => "Dry foods"
     ]);
-    $mGroup->recordManager->create([
+    $mGroup->record->create([
       "title" => "Liquids"
     ]);
 
     $mCountry = new Country($this->main);
 
-    $mSupplier->recordManager->create([
+    $mSupplier->record->create([
       "vat_id" => "GB123562563",
       "title" => "Fox Foods",
       "id_country" => $mCountry->eloquent->inRandomOrder()->first()->id,
     ]);
-    $mSupplier->recordManager->create([
+    $mSupplier->record->create([
       "vat_id" => "CZ123562563",
       "title" => "Bořek Furniture",
       "id_country" => $mCountry->eloquent->inRandomOrder()->first()->id,
     ]);
-    $mSupplier->recordManager->create([
+    $mSupplier->record->create([
       "vat_id" => "FR123562563",
       "title" => "Denise's Dry Goods",
       "id_country" => $mCountry->eloquent->inRandomOrder()->first()->id,
@@ -795,7 +795,7 @@ class GenerateDemoData extends \HubletoMain\Cli\Agent\Command
     ];
 
     foreach ($products as $product) {
-      $idProduct = $mProduct->recordManager->create([
+      $idProduct = $mProduct->record->create([
         "title" => $product[0],
         "unit_price" => $product[1],
         "margin" => $product[2],